move chart-related functions to chart.js
